Extract mock factory generator from play mock client
